Moved login service from a plugin back to libswganh
