+++ src/swganh/login/messages/login_cluster_status.h
@@ -17,8 +17,8 @@
  along with MMOServer.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef LOGIN_MESSAGES_LOGIN_CLUSTER_STATUS_H_
-#define LOGIN_MESSAGES_LOGIN_CLUSTER_STATUS_H_
+#ifndef SWGANH_LOGIN_MESSAGES_LOGIN_CLUSTER_STATUS_H_
+#define SWGANH_LOGIN_MESSAGES_LOGIN_CLUSTER_STATUS_H_
 
 #include <cstdint>
 #include <algorithm>
@@ -29,8 +29,9 @@
 #include "anh/byte_buffer.h"
 #include "swganh/base/swg_message.h"
 
-#include "login/galaxy_status.h"
+#include "swganh/login/galaxy_status.h"
 
+namespace swganh {
 namespace login {
 namespace messages {
 
@@ -90,6 +91,6 @@ struct LoginClusterStatus : public swganh::base::SwgMessage<LoginClusterStatus> 
 
 LoginClusterStatus BuildLoginClusterStatus(const std::vector<login::GalaxyStatus>& galaxy_status);
 
-}}  // namespace login::messages
+}}}  // namespace swganh::login::messages
 
-#endif  // LOGIN_MESSAGES_LOGIN_CLUSTER_STATUS_H_
+#endif  // SWGANH_LOGIN_MESSAGES_LOGIN_CLUSTER_STATUS_H_
